+++ src/main/java/com/assetflow/loan/Loan.java
@@ -18,6 +18,7 @@ public class Loan {
     @Column(name = "loan_id")
     private Long id;
 
+
     @Enumerated(EnumType.STRING)
     private LoanStatus loanStatus;
 
@@ -50,6 +51,9 @@ public class Loan {
         }
 
     }
+
+
+