+++ src/main/java/com/assetflow/loan/dto/LoanReturnResponse.java
@@ -1,0 +1,17 @@
+package com.assetflow.loan.dto;
+
+import com.assetflow.loan.LoanStatus;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Getter
+@AllArgsConstructor
+public class LoanReturnResponse {
+    private Long lonaId;
+    private LoanStatus loanStatus;
+    private Long memberId;
+    private Long assetItemId;
+    private LocalDate returnDate;
+}